refactor (global) : ErrorCode enum을 다형성을 활용해 도메인 결합도 최소화 (#236)
* feat (global) : ErrorCode enum을 다형성을 활용하여 도메인 결합도 낮춤

* feat (post) : ErrorCode enum을 다형성을 활용하여 도메인 결합도 낮춤

* feat (user) : ErrorCode enum을 다형성을 활용하여 도메인 결합도 낮춤

* test (post) : ErrorCode enum을 다형성을 활용하여 도메인 결합도 낮춤

* test (user) : ErrorCode enum을 다형성을 활용하여 도메인 결합도 낮춤
