+++ src/main/java/com/backend/immilog/post/application/CommentUploadServiceImpl.java
@@ -12,7 +12,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import static com.backend.immilog.global.exception.ErrorCode.POST_NOT_FOUND;
+import static com.backend.immilog.post.exception.PostErrorCode.POST_NOT_FOUND;
 
 @Slf4j
 @Service
