+++ src/main/java/com/backend/immilog/post/application/PostUploadServiceImpl.java
@@ -2,13 +2,13 @@ package com.backend.immilog.post.application;
 
 import com.backend.immilog.global.exception.CustomException;
 import com.backend.immilog.global.infrastructure.BulkInsertRepository;
-import com.backend.immilog.post.model.services.PostUploadService;
-import com.backend.immilog.post.model.repositories.PostRepository;
 import com.backend.immilog.post.model.entities.Post;
 import com.backend.immilog.post.model.entities.PostResource;
+import com.backend.immilog.post.model.repositories.PostRepository;
+import com.backend.immilog.post.model.services.PostUploadService;
 import com.backend.immilog.post.presentation.request.PostUploadRequest;
-import com.backend.immilog.user.model.interfaces.repositories.UserRepository;
 import com.backend.immilog.user.model.entities.User;
+import com.backend.immilog.user.model.interfaces.repositories.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -18,11 +18,11 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import static com.backend.immilog.global.exception.ErrorCode.FAILED_TO_SAVE_POST;
-import static com.backend.immilog.global.exception.ErrorCode.USER_NOT_FOUND;
 import static com.backend.immilog.post.enums.PostType.POST;
 import static com.backend.immilog.post.enums.ResourceType.ATTACHMENT;
 import static com.backend.immilog.post.enums.ResourceType.TAG;
+import static com.backend.immilog.post.exception.PostErrorCode.FAILED_TO_SAVE_POST;
+import static com.backend.immilog.user.exception.UserErrorCode.USER_NOT_FOUND;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -67,13 +67,13 @@ public class PostUploadServiceImpl implements PostUploadService {
         bulkInsertRepository.saveAll(
                 postResourceList,
                 """
-                        INSERT INTO post_resource (
-                            post_seq,
-                            post_type,
-                            resource_type,
-                            content
-                        ) VALUES (?, ?, ?, ?)
-                        """,
+                INSERT INTO post_resource (
+                    post_seq,
+                    post_type,
+                    resource_type,
+                    content
+                ) VALUES (?, ?, ?, ?)
+                """,
                 (ps, postResource) -> {
                     try {
                         ps.setLong(1, postResource.getPostSeq());
