+++ src/main/java/com/backend/immilog/post/enums/ReferenceType.java
@@ -1,11 +1,12 @@
 package com.backend.immilog.post.enums;
 
 import com.backend.immilog.global.exception.CustomException;
-import com.backend.immilog.global.exception.ErrorCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 import java.util.Arrays;
+
+import static com.backend.immilog.post.exception.PostErrorCode.INVALID_REFERENCE_TYPE;
 
 @Getter
 @RequiredArgsConstructor
@@ -21,6 +22,6 @@ public enum ReferenceType {
         return Arrays.stream(values())
                 .filter(type -> type.stringValue.equals(referenceType))
                 .findFirst()
-                .orElseThrow(() -> new CustomException(ErrorCode.INVALID_REFERENCE_TYPE));
+                .orElseThrow(() -> new CustomException(INVALID_REFERENCE_TYPE));
     }
 }
