+++ src/test/java/com/backend/immilog/user/application/UserInformationServiceTest.java
@@ -5,10 +5,10 @@ import com.backend.immilog.global.application.RedisService;
 import com.backend.immilog.global.exception.CustomException;
 import com.backend.immilog.global.security.JwtProvider;
 import com.backend.immilog.user.enums.UserStatus;
-import com.backend.immilog.user.model.interfaces.repositories.UserRepository;
 import com.backend.immilog.user.model.dtos.UserSignInDTO;
 import com.backend.immilog.user.model.embeddables.Location;
 import com.backend.immilog.user.model.entities.User;
+import com.backend.immilog.user.model.interfaces.repositories.UserRepository;
 import com.backend.immilog.user.presentation.request.UserInfoUpdateRequest;
 import com.backend.immilog.user.presentation.request.UserPasswordChangeRequest;
 import org.junit.jupiter.api.BeforeEach;
@@ -22,11 +22,11 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 
-import static com.backend.immilog.global.exception.ErrorCode.NOT_AN_ADMIN_USER;
-import static com.backend.immilog.global.exception.ErrorCode.PASSWORD_NOT_MATCH;
 import static com.backend.immilog.user.enums.Countries.*;
 import static com.backend.immilog.user.enums.UserRole.ROLE_ADMIN;
 import static com.backend.immilog.user.enums.UserRole.ROLE_USER;
+import static com.backend.immilog.user.exception.UserErrorCode.NOT_AN_ADMIN_USER;
+import static com.backend.immilog.user.exception.UserErrorCode.PASSWORD_NOT_MATCH;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.*;
